Move SIFT materials
Moving Sift stuff into its own folder
