+++ SIFT/OrientationHistogramUAV.py
