+++ SIFT/noise2.jpg
