+++ SIFT/noise3.jpg
